# LibGfx/JPEG: Reduce memory usage by almost 50% :^)