~Based on #17907 and #17786, only the 5 lasts commits are original.~

This is ready to merge!

I simplified this PR as the second part need more thought/work.

Master:
157,0 Mo
Completed benchmark 'small_image' on average in 0.0±0.0ms (min=0ms, max=0ms, total=0ms)
Completed benchmark 'big_image' on average in 614.2±7.5ms (min=597ms, max=645ms, total=61424ms)
Completed benchmark 'rgb_image' on average in 6.0±0.3ms (min=6ms, max=9ms, total=610ms)
Completed benchmark 'several_scans' on average in 5.2±0.5ms (min=5ms, max=7ms, total=528ms)


[LibGfx/JPEG: Use a smaller type to store coefficients](https://github.com/SerenityOS/serenity/pull/18077/commits/a28821b6307b0e8c8b841d278f6fe55d856f62cb):
84,8 Mo
Completed benchmark 'small_image' on average in 0.0±0.0ms (min=0ms, max=0ms, total=0ms)
Completed benchmark 'big_image' on average in 538.3±19.5ms (min=517ms, max=628ms, total=53835ms)
Completed benchmark 'rgb_image' on average in 6.0±0.1ms (min=6ms, max=7ms, total=604ms)
Completed benchmark 'several_scans' on average in 5.0±0.1ms (min=5ms, max=6ms, total=502ms)