@nico 2023-03-31
The actual change lgtm.

Notes on the benchmark:

1. Adding a 5.8 MiB binary file to the repo is maybe not ideal (?)
2. It's weird to me that the images are loaded at static initialization time instead of in the test itself. It doesn't even separate file system access from the test – all disk i/o will happen when the data is paged in on first access.

Neither of these are necessarily blocking (up to maintainers), but it seemed a bit surprising to me.