@LucasChollet 2023-03-31
> Adding a 5.8 MiB binary file to the repo is maybe not ideal (?)

I'm not very happy with that either but I decided to let a maintainer take the decision. However, I think that it is important to benchmark the decoder on bigger images.

> It's weird to me that the images are loaded at static initialization time instead of in the test itself. It doesn't even separate file system access from the test – all disk i/o will happen when the data is paged in on first access.

Oops, my bad mapped files were an error. Loading at static time was by design tho. I want to test the decoder regardless of I/O speed.