@nico 2023-04-03
Maybe consider pulling out the actual memory savings into their own PR – I'd imagine that part will be merged immediately :)