@nico 2023-04-03
(note to self: tests now in #18162)