@LucasChollet 2023-04-03
Hold on! The recently added 4th component isn't an i16.